requirements.txt
numpy==1.26.4
pandas==2.2.0
seaborn==0.13.2
matplotlib==3.8.3
yfinance==0.2.36
statsmodels==0.14.1
scipy==1.12.0
pylab-sdk==1.7.2